api/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
faiss-cpu==1.8.0.post1
pydantic==2.8.2
numpy==1.26.4
boto3==1.35.9